Added FlyingDusty flying + looking at player + always flying above certain height
Flying dusty is not kinematic, so Unity handles all physics stuff.

In each level there should be a MinFlyingHeightSetter object, which the enemies will find at their spawn and take its Y. That will be the minimum flying height. If they reach it, the next time they chose a direction, it will be mostly upwards.
OR you can just set their bool followPlayersHeight to true and every frame they will track the players height and use that as their MinFlyingHeight.

Enemies now FlyUpAndDown() the same way so just introduce sime offset of time after which they start to do this so they will not do it uniformly